pyproject.toml
[tool.poetry]
name = "gestion-salas"
version = "0.1.0"
description = "Gestor de salas de reunión"
authors = ["0xGeN02 <manuelmateodgl02@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.5"
uvicorn = {version = "^0.32.1", extras = ["standard"]}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
python-dotenv = "^1.0.1"
asyncpg = "^0.30.0"
pydantic-settings = "^2.6.1"
pydantic = {extras = ["email"], version = "^2.10.3"}
databases = "^0.9.0"
httpx = "^0.28.0"
greenlet = "^3.1.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
